+++ console-api/src/trace.rs
@@ -1,1 +1,3 @@
+#![allow(warnings)]
+
 include!("generated/rs.tokio.console.trace.rs");