+++ tokio-console/src/view/mod.rs
@@ -105,6 +105,16 @@ impl View {
 
         self.handl_help_popup(event);
 
+        if matches!(event, key!(Char('t'))) {
+            self.state = TasksList;
+            return update_kind;
+        }
+
+        if matches!(event, key!(Char('r'))) {
+            self.state = ResourcesList;
+            return update_kind;
+        }
+
         match self.state {
             TasksList => {
                 // The enter key changes views, so handle here since we can
@@ -119,9 +129,6 @@ impl View {
                             ));
                         }
                     }
-                    key!(Char('r')) => {
-                        self.state = ResourcesList;
-                    }
                     _ => {
                         // otherwise pass on to view
                         self.tasks_list.update_input(event);
@@ -135,9 +142,6 @@ impl View {
                             update_kind = UpdateKind::SelectResource(res.borrow().span_id());
                             self.state = ResourceInstance(self::resource::ResourceView::new(res));
                         }
-                    }
-                    key!(Char('t')) => {
-                        self.state = TasksList;
                     }
                     _ => {
                         // otherwise pass on to view