+++ tokio-console/src/view/styles.rs
@@ -47,12 +47,32 @@ impl Styles {
         }
     }
 
-    pub fn error_init(&self) -> color_eyre::Result<()> {
-        use color_eyre::config::{HookBuilder, Theme};
+    pub fn error_init(&self, cfg: &crate::config::Config) -> color_eyre::Result<()> {
+        use color_eyre::{
+            config::{HookBuilder, Theme},
+            ErrorKind,
+        };
 
         let mut builder = HookBuilder::new()
             .issue_url(concat!(env!("CARGO_PKG_REPOSITORY"), "/issues/new"))
+            .issue_filter(|kind| match kind {
+                // Only suggest reporting GitHub issues for panics, not for
+                // errors, so people don't open GitHub issues for stuff like not
+                // being able to find a config file or connections being
+                // terminated by remote hosts.
+                ErrorKind::NonRecoverable(_) => true,
+                ErrorKind::Recoverable(_) => false,
+            })
+            // filter out `color-eyre`'s default set of frames to skip from
+            // backtraces.
+            //
+            // this includes `std::rt`, `color_eyre`'s own frames, and
+            // `tokio::runtime` & friends.
+            .add_default_filters()
             .add_issue_metadata("version", env!("CARGO_PKG_VERSION"));
+        // Add all the config values to the GitHub issue metadata
+        builder = cfg.add_issue_metadata(builder);
+
         if self.palette == Palette::NoColors {
             // disable colors in error reports
             builder = builder.theme(Theme::new());