+++ console-api/CHANGELOG.md
@@ -1,4 +1,30 @@
-<a name=""></a>
+<a name="0.4.0"></a>
+## 0.4.0 (2022-08-10)
+
+
+#### Breaking Changes
+
+*  Update `tonic` to `0.8` (#364) ([40e2f6fd](40e2f6fd))
+
+#### Features
+
+*  Update `tonic` to `0.8` (#364) ([40e2f6fd](40e2f6fd))
+
+
+<a name="0.3.0"></a>
+## 0.3.0 (2022-05-23)
+
+
+#### Features
+
+*  add optional histogram outlier details (#351) ([46115913](46115913))
+
+#### Breaking Changes
+
+*   add optional histogram outlier details (#351) ([46115913](46115913))
+
+
+<a name="0.2.0"></a>
 ## 0.2.0 (2022-04-08)